--- a/flattener.go
+++ b/flattener.go
@@ -5,26 +5,21 @@ package quamina
 // values. The following example illustrates how it works for a JSON object:
 // { "a": 1, "b": "two", "c": true", "d": nil, "e": { "e1": 2, "e2":, 3.02e-5} "f": [33, "x"]} }
 // should produce
-//  "a", "1"
-//  "b", "\"two\"",
-//  "c", "true"
-//  "d", "nil",
-//  "e\ne1", "2"
-//  "e\ne2", "3.02e-5"
-//  "f", "33"
-//  "f", "\"x\""
+//
+//	"a", "1"
+//	"b", "\"two\"",
+//	"c", "true"
+//	"d", "nil",
+//	"e\ne1", "2"
+//	"e\ne2", "3.02e-5"
+//	"f", "33"
+//	"f", "\"x\""
 //
 // Let's call the first column, eg "d" and "e\ne1", the path. For each
-// step i the path, e.g. "d" and "e1", the Flattener should call
-// NameTracker.IsNameUsed(step) and if that comes back negative, not include any
-// paths which don't contain that step. So in the example above, if
-// nameTracker.IsNameUsed() only came back true for "a" and "f", then the output
-// would be
-//  "a", "1"
-//  "f", "33"
-//  "f", "\"x\""
+// step i the path, e.g. "d" and "e1", the Flattener should utilize SegmentsTreeTracker to
+// traverse the hierarchy and select only the needed fields.
 type Flattener interface {
-	Flatten(event []byte, tracker NameTracker) ([]Field, error)
+	Flatten(event []byte, tracker SegmentsTreeTracker) ([]Field, error)
 	Copy() Flattener
 }