--- a/matcher.go
+++ b/matcher.go
@@ -4,5 +4,5 @@ type matcher interface {
 	addPattern(x X, pat string) error
 	matchesForFields(fields []Field) ([]X, error)
 	deletePatterns(x X) error
-	IsNameUsed(label []byte) bool
+	getSegmentsTreeTracker() SegmentsTreeTracker
 }