## Context

Prior discussion is here: https://github.com/timbray/quamina/issues/113

Currently the flattener and more specifically "NameTracker" is not hierarchy aware, to explain that let's examine this case:
Pattern: `{ "context": { "user_id": [1] } }`
Event: `{ "context": { "user_id": 1 }, "payload": { /* lots of fields, large payload */ }`

In this case the flattener will read context and then user_id, all the other values (strings, objects and arrays) will be skipped, but still will get traversed. By making the "NameTracker" aware of the hierarchy, we can:
* Know that once we read "user_id" we can stop consuming the event and exit
* Cache the paths - so `pathForChild` is not needed and we reduce allocations.

## Hierarchy

To add Hierarchy to `NameTracker`, I introduced `SegmentsTreeTracker` interface and `segmentsTree` implementation, which is a simple tree that represents the living patterns, where:
* Node is an object / array we need to traverse into
* Field is leaf value - string/number/array we need to pluck as a Quamina's Field.

The current implementation and idea is to traverse the JSON together with `SegmentsTreeTracker` (you can think about it like a mirror), once we traverse them together we know:
* If the current field in his hierarchy (object/array) is needed - simple lookup on `fields`
* Get the path of a field - simple lookup, since `fields` is map from string (member name) to it's path (`[]byte`)
* Now if we can stop process the object/array - once we traverse all current nodes (objects/arrays) and fields - by decrementing a counter

## Benchmarks

```
name                                         old time/op    new time/op    delta
CityLots-10                                    4.42µs ± 0%    4.28µs ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattener_ContextFields-10                9.06µs ± 0%    0.21µs ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattener_MiddleNestedField-10            10.9µs ± 0%     1.8µs ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattener_LastField-10                    9.65µs ± 0%    9.68µs ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattner_Evaluate_ContextFields-10        9.42µs ± 0%    0.50µs ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattner_Evaluate_MiddleNestedField-10    11.0µs ± 0%     2.0µs ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattner_Evaluate_LastField-10            9.81µs ± 0%    9.87µs ± 0%   ~     (p=1.000 n=1+1)

name                                         old alloc/op   new alloc/op   delta
CityLots-10                                      985B ± 0%      832B ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattener_ContextFields-10                 48.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
_JsonFlattener_MiddleNestedField-10             64.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
_JsonFlattener_LastField-10                     32.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
_JsonFlattner_Evaluate_ContextFields-10          232B ± 0%      184B ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattner_Evaluate_MiddleNestedField-10      240B ± 0%       48B ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattner_Evaluate_LastField-10             80.0B ± 0%     48.0B ± 0%   ~     (p=1.000 n=1+1)

name                                         old allocs/op  new allocs/op  delta
CityLots-10                                      38.0 ± 0%      31.0 ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattener_ContextFields-10                  3.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
_JsonFlattener_MiddleNestedField-10              4.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
_JsonFlattener_LastField-10                      2.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
_JsonFlattner_Evaluate_ContextFields-10          9.00 ± 0%      6.00 ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattner_Evaluate_MiddleNestedField-10      9.00 ± 0%      3.00 ± 0%   ~     (p=1.000 n=1+1)
_JsonFlattner_Evaluate_LastField-10              5.00 ± 0%      3.00 ± 0%   ~     (p=1.000 n=1+1)
```

We can see the expected results:
* `CityLots` and `LastField` have improved only by allocations (for paths, no `pathForChild` calls), but speed isn't improved since we need to consume whole event here.
* `Context` & `MiddleNestedField` have improved by a lot since we can do early stop and we don't allocate for paths.

## What's next & Status

This PR is not ready to be merged, I am opening it to get an early feedback (given all tests passing, benchmark results) about the interface and implementation, once I get approval on that I can continue to:

- [x]  Adding tests: lots of tests are missing here - for paths indexing, objects as fields, etc..
- [ ]  Investigate degradation of `LastField`
- [x]  Improve the early stop case - currently it's funky with returning error and skipping object till the end, tests are needed here.
- [x]  Objects as fields - Currently this case is not covered, but assuming we have "exists: true" pattern on object, it shouldn't pass (didn't checked if we have test for it)
- [x]  Remove `NameTracker` and `namesUsed`

### Concern about skipping

We talked about this in the previous PR (https://github.com/timbray/quamina/pull/119), that we skip JSONs and now a bit more aggressive, and we can accept invalid JSON.

I have an idea on how to solve it - since there is a tradeoff about strictness and performance, I suggest let the user decide what he prefers with an option to flattener - if "strict mode" is enabled we won't do any skipping and no early exists.

If this is something we will to consider, I'll create separate issue and PR for it.