@embano1 2022-10-12
This is some great insights! On a related note, I posted a similar question/thought in the [`ruler`](https://github.com/aws/event-ruler/issues/46) project, which Tim was also involved. Not sure your findings translate 1:1, but this confirms my assumption of speeding up matching by the `context` (or `metadata`) use case, e.g. `CloudEvents` envelop.