@timbray 2022-10-12
BTW, I think the verb "pluck" is maybe not the best choice. In English it has two meanings which are opposite from each other. (English has way too many words like this.)

1. I plucked the best apple from the tree
2. I plucked the feathers from the chicken (i.e. removed them to throw away)

So the message is a bit confusing, do you pluck to select or pluck to discard?