@yosiat 2022-10-13
Thanks for the review!

> I think I agree with this but I'm not 100% sure I understand. Most of my comments here are minor but one is important: Could you put a long-ish explanatory comment on type HierarchicalNameTracker (and also maybe on the implementation in paths.go) explaining at a high level what this does. Lots of the central type implementations in Quamina have such a comment.

I will add such comment, but want to see if you accept the interface and the general idea.

>  I guess they will have to duplicate some of this work? In that spirit, is the code in paths.go specific to JSON?

The idea behind this interface and `paths.go` is to have a generic implementation that can be used for JSON and other future Flatteners. All of them are flattening hierarchal data, The difference might comes that ones won't need to traverse the hierarchy and he will a fast way to get value via path pointers, and this tree data structure might not be 100% efficient for them.

I think that we will need to challenge the Flattener with writing another one for data format which is different than JSON, I want to do it, but it's a good amount of work and time that I don't have right now, but I am planning to do it. Not sure when though.

> After you've written this, you may find you want to rename some of the types and methods… what is Get() getting? What is the Name in HierarchicalNameTracker? etc.

You wrote about this here and in the comments and I agree with you, I didn't maintained a consistent terminology - sometime it's a member, sometimes it's a property and sometimes it's a name.

Now that I understand (from you comments) the terminology in Quamina I'll make a names across the code in this PR.

Just one question for me to you, now that I understand that we have "Path" and it's parts are called "Segment", why you called it `NameTracker` and not `SegmentTracker` ?

> So the message is a bit confusing, do you pluck to select or pluck to discard?

I meant Pluck as Select and not discard :)

I used pluck because other languages and libraries have a method called `pluck` on object/hash which allows to select specific fields, like rails for example: https://apidock.com/rails/ActiveRecord/Calculations/pluck and in Ruby actually it's called select.

I'll rename pluck to select.