@yosiat 2022-10-13
@timbray 

Thanks for this amazing review, learned a lot, I'll add most of the comments, there are some that waiting for three things to be clarified/answered (and once those are clarified, I'll fix all comments):
1. `why you called it NameTracker and not SegmentTracker`
2. Do we want to keep `NameTracker` and `namesUsed` discussed here - https://github.com/timbray/quamina/pull/124#discussion_r994467358
3. Allowing to select objects as `quamina.Field` - https://github.com/timbray/quamina/pull/124#discussion_r994461942

Thanks 🙏