@timbray 2022-10-13
I agree, NameTracker certainly isn't the best name. Happy to accept something better. 

As I said, my impression is that making `"exists":true` work is nontrivial, but maybe you found a good way to do it and I didn't notice?  I don't want to hold up this nice optimization to wait for that if it's difficult.