@yosiat 2022-10-31
@timbray yes, I fixed most of the comments, added tests & fixed some.

The major thing we didn't answered about, is "NameTracker" & "namesUsed", do we still want to keep it? it's affecting both:
* Code coverage - since "IsNameUsed" is not used anymore
* The interface that flattener accepts, we had a discussion about this.

I am not sure where you are standing about, so we can do an elaborated discussion on that matter, but from my side:
* I don't believe in two solutions both "SegmentsTree" and "NameTracker", it will be bad for us to maintain and external flatteners to work with.
* I proved that "SegmentsTree" is a good solution for JSON flattening, but didn't prove it's good in general for flattening, I think it can be, but to actually validate it an additional flattener will need to be written (protobuf/cbor).