@timbray 2022-10-31
I 109% agree on refactoring and improving the names. Do you want to do that before I look again?