@yosiat 2022-10-31
@timbray I think you missed me If I read your comment correctly.

I did all the naming changes, the only thing left to decide is *can we remove NameTracker and replace it with SegmentsTree* ?