@timbray 2022-11-03
> @timbray The pr is ready for your re-review.

Excellent. Sorry, won't get there for a day or so.  Is there a good place to start, i.e. a big comment explaining the whole segmentsTree stuff?