@yosiat 2022-11-04
@timbray to review this PR I suggest:
1. Read the pull request body first
2. Read the interface & it's comment: segments_tree_tracker.go
3. Read the implementation - segments_tree.go (and it's tests to understand it's usage a bit)
4. Then it's first usage which is core_matcher in pattern addition
5. Lastly json flattener