@yosiat 2022-11-09
@timbray I pushed a rebase :)

BTW, I saw that benchmarks fails on regression in master (https://github.com/timbray/quamina/actions/runs/3430881550/jobs/5718364564), this PR will help with this regression.